refactor(schema): consistency naming file
