+++ app/admin/sections/SchemaSection/elements/index.ts
@@ -1,1 +1,2 @@
-export * from "./ManagementSchemaHeader";
+export * from "./schema-section-header";
+export * from "./schema-table";
