+++ app/admin/sections/SchemaSection/elements/schema-section-header.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-export const ManagementSchemaHeader = () => {
+export const SchemaSectionHeader = () => {
   return (
     <div className="w-1/2">
       <p className="font-semibold text-lg text-[#606060]">Manajemen Skema Database</p>
