+++ app/admin/sections/SchemaSection/index.tsx
@@ -1,10 +1,12 @@
 import React from "react";
-import { ManagementSchemaHeader } from "./elements";
+import { SchemaSectionHeader } from "./elements";
+import { SchemaTable } from "./elements/schema-table";
 
 const SchemaSection = () => {
   return (
-    <section className="container mx-auto max-w-screen-lg">
-      <ManagementSchemaHeader />
+    <section className="container mx-auto max-w-screen-lg flex flex-col gap-y-6">
+      <SchemaSectionHeader />
+      <SchemaTable />
     </section>
   );
 };
